split templates into individual files
